requirements.txt
Flask==2.2.2
torch==2.0.0
Werkzeug==2.2.2
torchvision==0.15.0
onnx==1.14.0
mlflow==2.5.0
pillow==9.5.0
numpy==1.23.4
onnxruntime==1.15.0